pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "risk-stratification-engine"
version = "0.1.0"
description = "Peterson-inspired athlete risk stratification research prototype"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "PyYAML>=6.0",
  "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pyarrow>=14",
]

[project.scripts]
risk-engine = "risk_stratification_engine.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
